+++ src/extension.ts
@@ -58,4 +58,7 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() { }
+export function deactivate() {
+    // Make sure that no process is running anymore after the windows was closed.
+    killAllProcesses();
+}